move IdentityServerConfig to database project
