+++ AspNetCore v5.0/Calabonga.Microservice.IdentityModule/Calabonga.Microservice.IdentityModule.Data/Calabonga.Microservice.IdentityModule.Data.csproj
@@ -8,6 +8,7 @@
 
   <ItemGroup>
     <PackageReference Include="Calabonga.UnitOfWork" Version="2.1.0" />
+    <PackageReference Include="IdentityServer4.EntityFramework" Version="4.1.2" />
     <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="5.0.4" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.0.4" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="5.0.4">
